restructure DataStore provider
